--- src/pages/FavouritesPage.tsx
@@ -1,80 +1,0 @@
-import { FC } from 'react';
-import { OfferData } from '../components/main-page/Card';
-import { OffersForCity } from '../components/favourites/OffersForCity';
-import { Link } from 'react-router-dom';
-
-export interface FavouritesPageProps {
-    offers: OfferData[]    
-}
-
-export const FavoritesPage: FC<FavouritesPageProps> = ({ offers }) => {
-
-  const offersByCities = Object.groupBy(offers, (item) => item.city)
-
-  return (
-  <div className="page">
-    <header className="header">
-      <div className="container">
-        <div className="header__wrapper">
-          <div className="header__left">
-            <Link className="header__logo-link" to="/">
-              <img
-                className="header__logo"
-                src="img/logo.svg"
-                alt="6 cities logo"
-                width="81"
-                height="41"
-              />
-            </Link>
-          </div>
-          <nav className="header__nav">
-            <ul className="header__nav-list">
-              <li className="header__nav-item user">
-                <a
-                  className="header__nav-link header__nav-link--profile"
-                  href="#"
-                >
-                  <div className="header__avatar-wrapper user__avatar-wrapper"></div>
-                  <span className="header__user-name user__name">
-                    Oliver.conner@gmail.com
-                  </span>
-                  <span className="header__favorite-count">3</span>
-                </a>
-              </li>
-              <li className="header__nav-item">
-                <a className="header__nav-link" href="#">
-                  <span className="header__signout">Sign out</span>
-                </a>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </div>
-    </header>
-
-    <main className="page__main page__main--favorites">
-      <div className="page__favorites-container container">
-        <section className="favorites">
-          <h1 className="favorites__title">Saved listing</h1>
-          <ul className="favorites__list">
-            {Object.keys(offersByCities).map(city => (
-              <OffersForCity city={city} offers={offersByCities[city] ?? null } key={city}/>
-            ))}
-          </ul>
-        </section>
-      </div>
-    </main>
-    <footer className="footer container">
-      <a className="footer__logo-link" href="main.html">
-        <img
-          className="footer__logo"
-          src="img/logo.svg"
-          alt="6 cities logo"
-          width="64"
-          height="33"
-        />
-      </a>
-    </footer>
-  </div>
-  )
-};
