+++ src/pages/offer/OfferPage.tsx
@@ -1,14 +1,18 @@
 import { FC } from 'react';
 import { Link, useParams } from 'react-router-dom';
-import { OFFERS_DATA } from '../mock/offers';
-import { SendReviewForm } from '../components/offer/SendReviewForm';
+import { OFFERS_DATA } from '../../mock/offers';
+import { SendReviewForm } from './components/SendReviewForm';
+import { FAVOURITES_URL } from '../../url';
 
 export const OfferPage: FC = () => {
-  let { id } = useParams();
-  const offerInfo = OFFERS_DATA.find(it => it.id + '' === id)
+  const { id } = useParams();
+  const offerInfo = OFFERS_DATA.find((it) => String(it.id) === id);
 
   if(!offerInfo) {
-    return <></>
+    return (
+      <>
+      </>
+    );
   }
 
   return (
@@ -30,16 +34,16 @@ export const OfferPage: FC = () => {
             <nav className="header__nav">
               <ul className="header__nav-list">
                 <li className="header__nav-item user">
-                  <a
+                  <Link
                     className="header__nav-link header__nav-link--profile"
-                    href="#"
+                    to={FAVOURITES_URL}
                   >
                     <div className="header__avatar-wrapper user__avatar-wrapper"></div>
                     <span className="header__user-name user__name">
                       Oliver.conner@gmail.com
                     </span>
                     <span className="header__favorite-count">3</span>
-                  </a>
+                  </Link>
                 </li>
                 <li className="header__nav-item">
                   <a className="header__nav-link" href="#">
@@ -377,5 +381,5 @@ export const OfferPage: FC = () => {
         </div>
       </main>
     </div>
-  )
-}
+  );
+};
