+++ src/url.ts
@@ -1,0 +1,4 @@
+export const FAVOURITES_URL = '/favourites';
+export const OFFER_URL = '/offer';
+export const MAIN_URL = '/main';
+export const LOGIN_URL = '/login';
